refactor: commands
