+++ src/main/java/net/ccbluex/liquidbounce/injection/mixins/minecraft/item/MixinItemStack.java
@@ -22,16 +22,18 @@ package net.ccbluex.liquidbounce.injection.mixins.minecraft.item;
 import net.ccbluex.liquidbounce.event.EventManager;
 import net.ccbluex.liquidbounce.event.events.ItemLoreQueryEvent;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.*;
 import net.minecraft.item.Item.TooltipContext;
+import net.minecraft.item.ItemStack;
 import net.minecraft.item.tooltip.TooltipType;
 import net.minecraft.text.Text;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Mixin(ItemStack.class)
 public class MixinItemStack {
@@ -39,8 +41,11 @@ public class MixinItemStack {
     @Inject(method = "getTooltip", at = @At("RETURN"), cancellable = true)
     void injectLoreQueryEvent(TooltipContext context, @Nullable PlayerEntity player, TooltipType type, CallbackInfoReturnable<List<Text>> cir) {
         List<Text> lore = cir.getReturnValue();
-        if (!(lore instanceof ArrayList<Text>)) return;
-        ItemLoreQueryEvent event = new ItemLoreQueryEvent((ItemStack) (Object) this, (ArrayList<Text>) lore);
+        if (!(lore instanceof ArrayList<Text>)) {
+            return;
+        }
+
+        ItemLoreQueryEvent event = new ItemLoreQueryEvent(ItemStack.class.cast(this), (ArrayList<Text>) lore);
         EventManager.INSTANCE.callEvent(event);
         cir.setReturnValue(event.getLore());
     }
